refactor: falled back on next.js pages router (the app router was not working)
