+++ src/app/layout.tsx
@@ -1,15 +1,5 @@
-import type { Metadata } from "next";
-import { Lexend } from "next/font/google";
-import { Analytics } from "@vercel/analytics/react";
-import { SpeedInsights } from "@vercel/speed-insights/next";
+import { Layout } from "@/components";
 import "@/styles/globals.css";
-
-const lexend = Lexend({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
 
 export default function RootLayout({
   children,
@@ -18,10 +8,8 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={lexend.className}>
-        {children}
-        <Analytics />
-        <SpeedInsights />
+      <body>
+        <Layout>{children}</Layout>
       </body>
     </html>
   );
