+++ src/components/Layout.tsx
@@ -1,0 +1,26 @@
+import type { Metadata } from "next";
+import { Lexend } from "next/font/google";
+import { Analytics } from "@vercel/analytics/react";
+import { SpeedInsights } from "@vercel/speed-insights/next";
+import "@/styles/globals.css";
+
+const lexend = Lexend({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className={lexend.className}>
+      {children}
+      <Analytics />
+      <SpeedInsights />
+    </div>
+  );
+}
